Added some casts (and some slightly different type decls) to quiet warnings in MSVC on Windows 10.
git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@8540 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ XSB/emu/incr_xsb.c
@@ -318,7 +318,7 @@ xsbBool incr_eval_builtin(CTXTdecl)
     sf  = get_call(CTXTc ptoc_tag(CTXTc 2), NULL);
     //    printf("sf1 %p\n",sf);
     if (IsNonNULL(sf)) {
-      ctop_int(CTXTc 3, (long) sf);
+      ctop_int(CTXTc 3, (Integer) sf);
       return TRUE;
     }
     else return FALSE;
